Rename rule_processor to operation to avoid confusions
