+++ src/collection.hpp
@@ -52,7 +52,7 @@ public:
         collection_cache &cache,
         const memory::unordered_map<ddwaf::rule *, exclusion::filter_mode> &rules_to_exclude,
         const memory::unordered_map<ddwaf::rule *, object_set> &objects_to_exclude,
-        const std::unordered_map<std::string, rule_processor::base::ptr> &dynamic_processors,
+        const std::unordered_map<std::string, operation::base::ptr> &dynamic_processors,
         ddwaf::timer &deadline) const;
 
 protected:
