+++ src/operation/ip_match.hpp
@@ -9,9 +9,9 @@
 #include <ip_utils.hpp>
 #include <memory>
 #include <radixlib.h>
-#include <rule_processor/base.hpp>
+#include <operation/base.hpp>
 
-namespace ddwaf::rule_processor {
+namespace ddwaf::operation {
 
 class ip_match : public base {
 public:
@@ -34,4 +34,4 @@ protected:
     std::unique_ptr<radix_tree_t, decltype(&radix_free)> rtree_{nullptr, nullptr};
 };
 
-} // namespace ddwaf::rule_processor
+} // namespace ddwaf::operation
