+++ src/operation/is_xss.hpp
@@ -7,18 +7,15 @@
 #pragma once
 
 #include <libinjection.h>
-#include <rule_processor/base.hpp>
+#include <operation/base.hpp>
 
-namespace ddwaf::rule_processor {
+namespace ddwaf::operation {
 
-class is_sqli : public base {
+class is_xss : public base {
 public:
-    is_sqli() = default;
-    [[nodiscard]] std::string_view name() const override { return "is_sqli"; }
+    is_xss() = default;
+    [[nodiscard]] std::string_view name() const override { return "is_xss"; }
     [[nodiscard]] std::optional<event::match> match(std::string_view pattern) const override;
-
-protected:
-    static constexpr unsigned fingerprint_length = 16;
 };
 
-} // namespace ddwaf::rule_processor
+} // namespace ddwaf::operation
