+++ src/ruleset.hpp
@@ -79,7 +79,7 @@ struct ruleset {
     std::unordered_map<std::string_view, exclusion::input_filter::ptr> input_filters;
 
     std::vector<rule::ptr> rules;
-    std::unordered_map<std::string, rule_processor::base::ptr> dynamic_processors;
+    std::unordered_map<std::string, operation::base::ptr> dynamic_processors;
 
     // The key used to organise collections is rule.type
     std::unordered_set<std::string_view> collection_types;
