+++ tests/operation/is_xss_test.cpp
@@ -6,50 +6,51 @@
 
 #include "../test.h"
 
-using namespace ddwaf::rule_processor;
+using namespace ddwaf::operation;
 
-TEST(TestIsSQLi, TestBasic)
+TEST(TestIsXSS, TestBasic)
 {
-    is_sqli processor;
+    is_xss processor;
     EXPECT_STREQ(processor.to_string().data(), "");
-    EXPECT_STREQ(processor.name().data(), "is_sqli");
+    EXPECT_STREQ(processor.name().data(), "is_xss");
 
     ddwaf_object param;
-    ddwaf_object_string(&param, "'OR 1=1/*");
+    ddwaf_object_string(&param, "<script>alert(1);</script>");
 
     auto match = processor.match_object(&param);
-    EXPECT_TRUE(match);
-    EXPECT_STREQ(match->resolved.c_str(), "'OR 1=1/*");
+    EXPECT_TRUE(match.has_value());
+
+    EXPECT_STREQ(match->resolved.c_str(), "<script>alert(1);</script>");
 
     ddwaf_object_free(&param);
 }
 
-TEST(TestIsSQLi, TestNoMatch)
+TEST(TestIsXSS, TestNoMatch)
 {
-    is_sqli processor;
+    is_xss processor;
 
     ddwaf_object param;
-    ddwaf_object_string(&param, "*");
+    ddwaf_object_string(&param, "non-xss");
 
     EXPECT_FALSE(processor.match_object(&param));
 
     ddwaf_object_free(&param);
 }
 
-TEST(TestIsSQLi, TestInvalidInput)
+TEST(TestIsXSS, TestInvalidInput)
 {
-    is_sqli processor;
+    is_xss processor;
 
     EXPECT_FALSE(processor.match({nullptr, 0}));
     EXPECT_FALSE(processor.match({nullptr, 30}));
-    EXPECT_FALSE(processor.match({"*", 0}));
+    EXPECT_FALSE(processor.match({"non-xss", 0}));
 }
 
-TEST(TestIsSQLi, TestRuleset)
+TEST(TestIsXSS, TestRuleset)
 {
     // Initialize a PowerWAF rule
     auto rule = readRule(
-        R"({version: '2.1', rules: [{id: 1, name: rule1, tags: {type: flow1, category: category1}, conditions: [{operator: is_sqli, parameters: {inputs: [{address: arg1}]}}]}]})");
+        R"({version: '2.1', rules: [{id: 1, name: rule1, tags: {type: flow1, category: category1}, conditions: [{operator: is_xss, parameters: {inputs: [{address: arg1}]}}]}]})");
     ASSERT_TRUE(rule.type != DDWAF_OBJ_INVALID);
 
     ddwaf_handle handle = ddwaf_init(&rule, nullptr, nullptr);
@@ -61,19 +62,21 @@ TEST(TestIsSQLi, TestRuleset)
 
     ddwaf_object param, tmp;
     ddwaf_object_map(&param);
-    ddwaf_object_map_add(&param, "arg1", ddwaf_object_string(&tmp, "'OR 1=1/*"));
+    ddwaf_object_map_add(&param, "arg1", ddwaf_object_string(&tmp, "<script>alert(1);</script>"));
 
     ddwaf_result ret;
 
     auto code = ddwaf_run(context, &param, &ret, LONG_TIME);
     EXPECT_EQ(code, DDWAF_MATCH);
     EXPECT_FALSE(ret.timeout);
-    EXPECT_EVENTS(ret,
-        {.id = "1",
-            .name = "rule1",
-            .tags = {{"type", "flow1"}, {"category", "category1"}},
-            .matches = {
-                {.op = "is_sqli", .address = "arg1", .value = "'OR 1=1/*", .highlight = "s&1c"}}});
+    EXPECT_EVENTS(ret, {.id = "1",
+                           .name = "rule1",
+                           .tags = {{"type", "flow1"}, {"category", "category1"}},
+                           .matches = {{
+                               .op = "is_xss",
+                               .address = "arg1",
+                               .value = "<script>alert(1);</script>",
+                           }}});
     ddwaf_result_free(&ret);
 
     ddwaf_context_destroy(context);
