+++ tests/rule_test.cpp
@@ -15,7 +15,7 @@ TEST(TestRule, Match)
     targets.push_back({get_target_index("http.client_ip"), "http.client_ip", {}, {}});
 
     auto cond = std::make_shared<condition>(std::move(targets),
-        std::make_unique<rule_processor::ip_match>(std::vector<std::string_view>{"192.168.0.1"}));
+        std::make_unique<operation::ip_match>(std::vector<std::string_view>{"192.168.0.1"}));
 
     std::vector<std::shared_ptr<condition>> conditions{std::move(cond)};
 
@@ -60,7 +60,7 @@ TEST(TestRule, NoMatch)
     targets.push_back({get_target_index("http.client_ip"), "http.client_ip", {}, {}});
 
     auto cond = std::make_shared<condition>(std::move(targets),
-        std::make_unique<rule_processor::ip_match>(std::vector<std::string_view>{}));
+        std::make_unique<operation::ip_match>(std::vector<std::string_view>{}));
 
     std::vector<std::shared_ptr<condition>> conditions{std::move(cond)};
     std::unordered_map<std::string, std::string> tags{{"type", "type"}, {"category", "category"}};
@@ -88,7 +88,7 @@ TEST(TestRule, ValidateCachedMatch)
         std::vector<condition::target_type> targets;
         targets.push_back({get_target_index("http.client_ip"), "http.client_ip", {}, {}});
         auto cond = std::make_shared<condition>(
-            std::move(targets), std::make_unique<rule_processor::ip_match>(
+            std::move(targets), std::make_unique<operation::ip_match>(
                                     std::vector<std::string_view>{"192.168.0.1"}));
         conditions.push_back(std::move(cond));
     }
@@ -97,7 +97,7 @@ TEST(TestRule, ValidateCachedMatch)
         std::vector<condition::target_type> targets;
         targets.push_back({get_target_index("usr.id"), "usr.id", {}, {}});
         auto cond = std::make_shared<condition>(std::move(targets),
-            std::make_unique<rule_processor::exact_match>(std::vector<std::string>{"admin"}));
+            std::make_unique<operation::exact_match>(std::vector<std::string>{"admin"}));
         conditions.push_back(std::move(cond));
     }
 
@@ -169,7 +169,7 @@ TEST(TestRule, MatchWithoutCache)
         std::vector<condition::target_type> targets;
         targets.push_back({get_target_index("http.client_ip"), "http.client_ip", {}, {}});
         auto cond = std::make_shared<condition>(
-            std::move(targets), std::make_unique<rule_processor::ip_match>(
+            std::move(targets), std::make_unique<operation::ip_match>(
                                     std::vector<std::string_view>{"192.168.0.1"}));
         conditions.push_back(std::move(cond));
     }
@@ -178,7 +178,7 @@ TEST(TestRule, MatchWithoutCache)
         std::vector<condition::target_type> targets;
         targets.push_back({get_target_index("usr.id"), "usr.id", {}, {}});
         auto cond = std::make_shared<condition>(std::move(targets),
-            std::make_unique<rule_processor::exact_match>(std::vector<std::string>{"admin"}));
+            std::make_unique<operation::exact_match>(std::vector<std::string>{"admin"}));
         conditions.push_back(std::move(cond));
     }
 
@@ -244,7 +244,7 @@ TEST(TestRule, NoMatchWithoutCache)
         std::vector<condition::target_type> targets;
         targets.push_back({get_target_index("http.client_ip"), "http.client_ip", {}, {}});
         auto cond = std::make_shared<condition>(
-            std::move(targets), std::make_unique<rule_processor::ip_match>(
+            std::move(targets), std::make_unique<operation::ip_match>(
                                     std::vector<std::string_view>{"192.168.0.1"}));
         conditions.push_back(std::move(cond));
     }
@@ -253,7 +253,7 @@ TEST(TestRule, NoMatchWithoutCache)
         std::vector<condition::target_type> targets;
         targets.push_back({get_target_index("usr.id"), "usr.id", {}, {}});
         auto cond = std::make_shared<condition>(std::move(targets),
-            std::make_unique<rule_processor::exact_match>(std::vector<std::string>{"admin"}));
+            std::make_unique<operation::exact_match>(std::vector<std::string>{"admin"}));
         conditions.push_back(std::move(cond));
     }
 
@@ -300,7 +300,7 @@ TEST(TestRule, FullCachedMatchSecondRun)
         std::vector<condition::target_type> targets;
         targets.push_back({get_target_index("http.client_ip"), "http.client_ip", {}, {}});
         auto cond = std::make_shared<condition>(
-            std::move(targets), std::make_unique<rule_processor::ip_match>(
+            std::move(targets), std::make_unique<operation::ip_match>(
                                     std::vector<std::string_view>{"192.168.0.1"}));
         conditions.push_back(std::move(cond));
     }
@@ -309,7 +309,7 @@ TEST(TestRule, FullCachedMatchSecondRun)
         std::vector<condition::target_type> targets;
         targets.push_back({get_target_index("usr.id"), "usr.id", {}, {}});
         auto cond = std::make_shared<condition>(std::move(targets),
-            std::make_unique<rule_processor::exact_match>(std::vector<std::string>{"admin"}));
+            std::make_unique<operation::exact_match>(std::vector<std::string>{"admin"}));
         conditions.push_back(std::move(cond));
     }
 
@@ -357,7 +357,7 @@ TEST(TestRule, ExcludeObject)
     targets.push_back({get_target_index("http.client_ip"), "http.client_ip", {}, {}});
 
     auto cond = std::make_shared<condition>(std::move(targets),
-        std::make_unique<rule_processor::ip_match>(std::vector<std::string_view>{"192.168.0.1"}));
+        std::make_unique<operation::ip_match>(std::vector<std::string_view>{"192.168.0.1"}));
 
     std::vector<std::shared_ptr<condition>> conditions{std::move(cond)};
     std::unordered_map<std::string, std::string> tags{{"type", "type"}, {"category", "category"}};
